IsPerfectMatch implemented.

+++ Core/Application/Services/IYarnAlternativeService.cs
@@ -7,5 +7,5 @@ public interface IYarnAlternativeService
 {
     Task<Result> GetYarnAlternatives(FindAlternativeSameNeedleOneThread request, CancellationToken cancellationToken = default);
 
-    Task<IEnumerable<YarnSuggestion>> FindSingleYarnSuggestions(double targetGauge, double targetNeedle);
+    Task<IEnumerable<YarnSuggestion>> FindSingleYarnSuggestions(int targetGauge, double targetNeedle);
 }
